refactor exhibitions component
